requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
Pillow==12.1.1
pytesseract==0.3.10
python-dotenv==1.0.0
PyPDF2==3.0.1
flask-cors==4.0.0
validators==0.22.0
tldextract==5.1.0
openai==0.28.0
gunicorn==21.2.0